Initial commit
So far, clearing the data table is implemented (not yet used),
fully reading the brainfuck code and tracking its length, as
well as validating a code symbol and calling a respective handling
function. Furthermore, a custom stack is implemented that holds
table entries for unclosed brackets during parsing. Additionally,
a pointer for keeping track of the currently inspected table entry
is also implemented.

Popping open bracket when the corresponding closing bracket is read
and actually using the bracket jump table produced by parsing to
interpret the brainfuck code is missing.