Added BCD conversion for indices
With this change, it is possible to calculate the BCD representation of
the DPTR into R0, R1 and R2. That can then be used to print the index of
an error to the LCD.

It basically works, but adjustments must be made regarding code style
and the index being display when a bracket is mismatched. First, I
want to diplay the index of the opened unbalanced bracket and secondly
the index of a superfluous closed bracket.