Added index for unclosed bracket error message
The trick is to get the symbol index of the topmost unclosed
bracket, as it is the one causing the parsing problem.
First, the correct memory bank must be selected and the XSTACK
pointer must be loaded. Then, the symbol pointer can be loaded,
ORG changes can be reverted. Lastly, the DPTR is converted to BCD.