Fixed faulty TPTR initialization
It came to be due to the fact, that I recently included the ORG
directive, which shifts the DPTR for reading symbols. I accounted
for that, by letting the DPTR start at the given offset when
interpreting. But then, I read the DPTR value as TPTR start,
because I used a cheap trick, that would directly put zero into
the TPTR from the time, when the DPTR would still start at zero.